Added javadoc to all files and ensure all packages and classes following cs2103t coding standard
