--- src/main/java/Exceptions/ChatResponseException/ChatResponseException.java
@@ -1,9 +1,0 @@
-package Exceptions.ChatResponseException;
-
-public class ChatResponseException extends Exception {
-
-    @Override
-    public String toString() {
-        return "Invalid command. Check the manual again Manager.";
-    }
-}
