--- src/main/java/Exceptions/PrintListException/EmptyListException.java
@@ -1,8 +1,0 @@
-package Exceptions.PrintListException;
-
-public class EmptyListException extends PrintListException {
-    @Override
-    public String toString() {
-        return "No data has been stored in the database.";
-    }
-}
