--- src/main/java/Exceptions/TaskCreationException/EmptyDetailException.java
@@ -1,9 +1,0 @@
-package Exceptions.TaskCreationException;
-
-public class EmptyDetailException extends TaskCreationException {
-
-    @Override
-    public String toString() {
-        return "Task description cannot be empty.";
-    }
-}
