--- src/main/java/Exceptions/TaskModificationException/TaskModificationException.java
@@ -1,9 +1,0 @@
-package Exceptions.TaskModificationException;
-
-public class TaskModificationException extends Exception {
-
-    @Override
-    public String toString() {
-        return "Error modifying the list.";
-    }
-}
