--- src/main/java/Exceptions/TaskModificationException/WrongSyntaxException.java
@@ -1,9 +1,0 @@
-package Exceptions.TaskModificationException;
-
-public class WrongSyntaxException extends TaskModificationException {
-
-    @Override
-    public String toString() {
-        return "Enter the rank of the list item you want to modify. Negative number is not accepted.";
-    }
-}
